requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
sounddevice>=0.4.0
soundfile>=0.12.0
edge-tts>=6.1.0
numpy>=1.24.0
pydub>=0.25.0
elevenlabs>=1.0.0
